+++ src/main/java/me/ashif/repository/PurchaseRepository.java
@@ -1,7 +1,6 @@
 package me.ashif.repository;
 
 
-import me.ashif.model.PurchaseInvoiceModel;
 import me.ashif.model.PurchaseModel;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -12,6 +11,6 @@ import java.util.List;
  * Created by techjini on 9/1/17.
  */
 @Repository
-public interface PurchaseRepository extends CrudRepository<PurchaseModel,Long> {
+public interface PurchaseRepository extends CrudRepository<PurchaseModel,Integer> {
     List<PurchaseModel> findBysupplierName(String supplierName);
 }
